--- a/optuna/_gp/optim_mixed.py
+++ b/optuna/_gp/optim_mixed.py
@@ -4,7 +4,10 @@ import math
 from typing import TYPE_CHECKING
 
 import numpy as np
+import scipy
+import torch
 
+from optuna._gp.batched_lbfgsb import batched_lbfgsb
 from optuna._gp.scipy_blas_thread_patch import single_blas_thread_if_scipy_v1_15_or_newer
 from optuna.logging import get_logger
 
@@ -20,14 +23,18 @@ else:
 _logger = get_logger(__name__)
 
 
-def _gradient_ascent(
+def is_scipy_version_supported() -> bool:
+    return scipy.__version__ <= "1.15.0"
+
+
+def _gradient_ascent_batched(
     acqf: BaseAcquisitionFunc,
-    initial_params: np.ndarray,
-    initial_fval: float,
+    initial_params_batched: np.ndarray,
+    initial_fvals: np.ndarray,
     continuous_indices: np.ndarray,
     lengthscales: np.ndarray,
     tol: float,
-) -> tuple[np.ndarray, float, bool]:
+) -> tuple[np.ndarray, np.ndarray, bool]:
     """
     This function optimizes the acquisition function using preconditioning.
     Preconditioning equalizes the variances caused by each parameter and
@@ -44,32 +51,68 @@ def _gradient_ascent(
     As the domain of `x` is [0, 1], that of `z` becomes [0, 1/l].
     """
     if len(continuous_indices) == 0:
-        return initial_params, initial_fval, False
-    normalized_params = initial_params.copy()
+        return initial_params_batched, initial_fvals, False
+    normalized_params = initial_params_batched.copy()
 
-    def negative_acqf_with_grad(scaled_x: np.ndarray) -> tuple[float, np.ndarray]:
+    # scaled_x: (B,D) or (D,)
+    def negative_acqf_with_grad(
+        scaled_x: np.ndarray, unconverged_batch_indices: np.ndarray
+    ) -> tuple[np.ndarray, np.ndarray]:
         # Scale back to the original domain, i.e. [0, 1], from [0, 1/s].
-        normalized_params[continuous_indices] = scaled_x * lengthscales
-        (fval, grad) = acqf.eval_acqf_with_grad(normalized_params)
+        assert scaled_x.ndim == 2
+        normalized_params[np.ix_(unconverged_batch_indices, continuous_indices)] = (
+            scaled_x * lengthscales
+        )
+        x_tensor = torch.from_numpy(
+            normalized_params[unconverged_batch_indices, :]
+        ).requires_grad_(True)
+        neg_fvals = -acqf.eval_acqf(x_tensor)
+        neg_fvals.sum().backward()
+        grads = x_tensor.grad.detach().numpy()  # type: ignore
+        neg_fvals = neg_fvals.detach().numpy()
         # Flip sign because scipy minimizes functions.
         # Let the scaled acqf be g(x) and the acqf be f(sx), then dg/dx = df/dx * s.
-        return -fval, -grad[continuous_indices] * lengthscales
+        return neg_fvals, grads[:, continuous_indices] * lengthscales
+
+    def _1D_wrapper(scaled_x: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
+        # 1D wrapper for the negative acquisition function with gradient.
+        unconverged_batch_indices = [0]
+        fval, grad = negative_acqf_with_grad(scaled_x[None], unconverged_batch_indices)
+        return fval.item(), grad.ravel()
 
     with single_blas_thread_if_scipy_v1_15_or_newer():
-        scaled_cont_x_opt, neg_fval_opt, info = so.fmin_l_bfgs_b(
-            func=negative_acqf_with_grad,
-            x0=normalized_params[continuous_indices] / lengthscales,
-            bounds=[(0, 1 / s) for s in lengthscales],
-            pgtol=math.sqrt(tol),
-            maxiter=200,
-        )
+        x0_batched = normalized_params[:, continuous_indices] / lengthscales
+        bounds = np.array([(0, 1 / s) for s in lengthscales])
+        pgtol = math.sqrt(tol)
+        max_iters = 200
+        # if is_scipy_version_supported():
+        if False:
+            scaled_cont_x_opts, neg_fval_opts, _ = batched_lbfgsb(
+                func_and_grad=negative_acqf_with_grad,
+                x0=x0_batched,
+                bounds=bounds,
+                pgtol=pgtol,
+                max_iters=max_iters,
+            )
+        else:
+            # TODO
+            scaled_cont_x_opts, neg_fval_opts = [], []
+            for x0 in x0_batched:
+                scaled_cont_x_opt, neg_fval_opt, _ = so.fmin_l_bfgs_b(
+                    func=_1D_wrapper,
+                    x0=x0,
+                    bounds=bounds,
+                    pgtol=pgtol,
+                    maxiter=max_iters,
+                )
+                scaled_cont_x_opts.append(scaled_cont_x_opt)
+                neg_fval_opts.append(neg_fval_opt)
+            scaled_cont_x_opts = np.array(scaled_cont_x_opts)
+            neg_fval_opts = np.array(neg_fval_opts)
+    normalized_params[:, continuous_indices] = scaled_cont_x_opts * lengthscales
 
-    if -neg_fval_opt > initial_fval and info["nit"] > 0:  # Improved.
-        # `nit` is the number of iterations.
-        normalized_params[continuous_indices] = scaled_cont_x_opt * lengthscales
-        return normalized_params, -neg_fval_opt, True
-
-    return initial_params, initial_fval, False  # No improvement.
+    # TODO: return update information for each parameter
+    return normalized_params, -neg_fval_opts, True
 
 
 def _exhaustive_search(
@@ -179,15 +222,39 @@ def _local_search_discrete(
         return _discrete_line_search(acqf, initial_params, initial_fval, param_idx, choices, xtol)
 
 
-def local_search_mixed(
+def _local_search_discrete_batched(
     acqf: BaseAcquisitionFunc,
-    initial_normalized_params: np.ndarray,
+    initial_params_batched: np.ndarray,
+    initial_fvals: np.ndarray,
+    param_idx: int,
+    choices: np.ndarray,
+    xtol: float,
+    last_changed_param,
+) -> tuple[np.ndarray, np.ndarray, bool]:
+    best_normalized_params_batched = initial_params_batched.copy()
+    best_fvals = initial_fvals.copy()
+    for batch, normalized_params in enumerate(initial_params_batched):
+        if last_changed_param == param_idx:
+            # Parameters not changed since last time.
+            best_fvals[batch] = float(acqf.eval_acqf_no_grad(normalized_params))
+        (best_normalized_params, best_fval, updated) = _local_search_discrete(
+            acqf, normalized_params, best_fvals[batch], param_idx, choices, xtol
+        )
+        best_normalized_params_batched[batch] = best_normalized_params
+        best_fvals[batch] = best_fval
+
+    # TODO: return updated properly
+    return best_normalized_params_batched, best_fvals, True
+
+
+def local_search_mixed_batched(
+    acqf: BaseAcquisitionFunc,
+    initial_normalized_params_batched: np.ndarray,
     *,
     tol: float = 1e-4,
-    max_iter: int = 100,
-) -> tuple[np.ndarray, float]:
+    max_iter: int = 10,
+) -> tuple[np.ndarray, np.ndarray]:
     continuous_indices = acqf.search_space.continuous_indices
-
     # This is a technique for speeding up optimization.
     # We use an isotropic kernel, so scaling the gradient will make
     # the hessian better-conditioned.
@@ -195,7 +262,6 @@ def local_search_mixed(
     # but for simplicity, the ones from the objective function are being reused.
     # TODO(kAIto47802): Think of a better way to handle this.
     lengthscales = acqf.length_scales[continuous_indices]
-
     choices_of_discrete_params = acqf.search_space.get_choices_of_discrete_params()
 
     discrete_xtols = [
@@ -205,45 +271,48 @@ def local_search_mixed(
         for choices in choices_of_discrete_params
     ]
 
-    best_normalized_params = initial_normalized_params.copy()
-    best_fval = float(acqf.eval_acqf_no_grad(best_normalized_params))
-
-    CONTINUOUS = -1
-    last_changed_param: int | None = None
+    best_normalized_params_batched = initial_normalized_params_batched.copy()
+    best_fvals = np.array(
+        [float(acqf.eval_acqf_no_grad(p)) for p in best_normalized_params_batched]
+    )
 
     for _ in range(max_iter):
-        if last_changed_param == CONTINUOUS:
-            # Parameters not changed since last time.
-            return best_normalized_params, best_fval
-        (best_normalized_params, best_fval, updated) = _gradient_ascent(
+        (best_normalized_params_batched, best_fvals, updated) = _gradient_ascent_batched(
             acqf,
-            best_normalized_params,
-            best_fval,
+            best_normalized_params_batched,
+            best_fvals,
             continuous_indices,
             lengthscales,
             tol,
         )
+
+        CONTINUOUS = -1
+        last_changed_param: int | None = None
         if updated:
             last_changed_param = CONTINUOUS
+        best_fvals = np.array(acqf.eval_acqf_no_grad(best_normalized_params_batched))
 
         for i, choices, xtol in zip(
             acqf.search_space.discrete_indices, choices_of_discrete_params, discrete_xtols
         ):
-            if last_changed_param == i:
-                # Parameters not changed since last time.
-                return best_normalized_params, best_fval
-            (best_normalized_params, best_fval, updated) = _local_search_discrete(
-                acqf, best_normalized_params, best_fval, i, choices, xtol
+            (best_normalized_params_batched, best_fvals, updated) = _local_search_discrete_batched(
+                acqf,
+                best_normalized_params_batched,
+                best_fvals,
+                i,
+                choices,
+                xtol,
+                last_changed_param,
             )
             if updated:
                 last_changed_param = i
 
-        if last_changed_param is None:
-            # Parameters not changed from the beginning.
-            return best_normalized_params, best_fval
-
-    _logger.warning("local_search_mixed: Local search did not converge.")
-    return best_normalized_params, best_fval
+            # if last_changed_param is None:
+            #     best_normalized_params_batched[batch] = best_normalized_params
+            #     best_fvals[batch] = float(acqf.eval_acqf_no_grad(best_normalized_params))
+        # TODO: Implement logging
+        # _logger.warning("local_search_mixed: Local search did not converge.")
+    return best_normalized_params_batched, best_fvals
 
 
 def optimize_acqf_mixed(
@@ -295,10 +364,10 @@ def optimize_acqf_mixed(
     best_x = sampled_xs[max_i, :]
     best_f = float(f_vals[max_i])
 
-    for x_warmstart in np.vstack([sampled_xs[chosen_idxs, :], warmstart_normalized_params_array]):
-        x, f = local_search_mixed(acqf, x_warmstart, tol=tol)
+    x_warmstarts = np.vstack([sampled_xs[chosen_idxs, :], warmstart_normalized_params_array])
+    xs, fs = local_search_mixed_batched(acqf, x_warmstarts, tol=tol)
+    for x, f in zip(xs, fs):
         if f > best_f:
             best_x = x
             best_f = f
-
     return best_x, best_f