## 概要 (Summary)

本PRは、GPSampler における連続パラメータの獲得関数の最適化処理を高速化するため、バッチ評価を導入します。これにより、特に連続値パラメータを多く含む目的関数において、最適化の実行時間を大幅に短縮することを目指します。

## 提案手法 (Solution)

連続値パラメータの最適化処理を複数同時に実行する「バッチ評価」機能を実装しました。

- **バッチ評価 (Batched Evaluation)**: 複数のTrialにまたがる連続値パラメータの最適化計算をまとめて実行します。これにより、計算リソースを効率的に利用し、処理時間を短縮します。ただし、SciPyの内部モジュールに依存しているため、問題があれば従来の逐次実行にもすぐに切り替えできるようにします。（フォールバック）
- **フォールバック (Fallback)**: SciPyの内部モジュールのAPIが変更した場合など、バッチ評価の前提条件が満たされない場合は、安全に従来の逐次評価処理へ切り替わります。

これらの切り替えは、`optim_mixed.py`の`_gradient_ascent_batched`関数の内部の条件分岐で切り替わります。こちらの条件分岐を適宜変える（`if False: ...`にするなど）ことで、簡単にバッチ評価から逐次評価に切り替えることができます。

```python

def _gradient_ascent_batched(...):
	...
	# SciPyの内部モジュールが利用可能な場合、バッチ評価を行う
	if is_scipy_version_supported():
		scaled_cont_x_opts, neg_fval_opts, info = batched_lbfgsb(...)
		...


	else: # Fallback
		...
		for batch, x0 in enumerate(x0_batched):
			...
			scaled_cont_x_opt, neg_fval_opt, info = so.fmin_l_bfgs_b(...)
			...
```

現状の切り替えは、`is_scipy_version_supported`関数にて切り替えの判定を行なっており、これは次のように定義されています。

```python
def is_scipy_version_supported() -> bool:
    return scipy.__version__ <= "1.15.0"
```

つまりSciPyのバージョンが`1.15.0`以下ならバッチ評価を、そうでなければ逐次評価を行うということです。

## 実験結果 (Experimental Results)

提案手法の有効性を確認するため、以下の2つの実験を行いました。
- `Original`: masterブランチの挙動
- `This PR (Batched)`: 提案手法のバッチ評価が機能した場合
- `This PR (Fallback)`: 提案手法がフォールバックした場合

**実験設定**
- `n_trials`: 100
- `seed`: 42
- `CPU`: M4 Mac

---

### 実験1: Rotated Ellipsoid + 離散変数

連続変数と離散変数が混在する問題設定です。

**目的関数**
```python
import numpy as np

NUM_CONTINUOUS_PARAMS = 3
# rotation matrix
ROTATION_MATRIX, _ = np.linalg.qr(np.random.rand(NUM_CONTINUOUS_PARAMS, NUM_CONTINUOUS_PARAMS))

def objective(trial):
    xs = np.array([trial.suggest_float(f"x_{i}", -1, 1) for i in range(NUM_CONTINUOUS_PARAMS)])
    xs = ROTATION_MATRIX @ xs
    z = trial.suggest_int("z", 0, 5)
    w = trial.suggest_int("w", 0, 2)
    return np.sum(xs**2) + (z - 2) ** 2 + (w - 1) ** 2
```

#### 結果1. 実行時間と目的関数の最適値

| stats\mode   | This PR (Batched) | This PR (Fallback) | Original    |
| ------------ | ----------------- | ------------------ | ----------- |
| Runtime      | 4.24 sec          | 5.47 sec           | 5.46 sec    |
| `best_value` | `8.491e-06`       | `5.264e-06`        | `1.173e-05` |

#### 結果2. Trialの初期では、どの手法も最適解がほとんど一致する

| Trial | Mode               | x0                   | x1                  | x2                   | z   | w   |
| ----- | ------------------ | -------------------- | ------------------- | -------------------- | --- | --- |
| 9     | Original           | 0.32504456870796394  | -0.3765778478211781 | 0.040136042355621626 | 3   | 0   |
| 9     | This PR (Batched)  | 0.32504456870796394  | -0.3765778478211781 | 0.040136042355621626 | 3   | 0   |
| 9     | This PR (Fallback) | 0.32504456870796394  | -0.3765778478211781 | 0.040136042355621626 | 3   | 0   |
| 10    | Original           | 0.04939102872817558  | 0.5183390069820595  | -1.0                 | 2   | 2   |
| 10    | This PR (Batched)  | 0.04939102872819379  | 0.5183390069820168  | -1.0                 | 2   | 2   |
| 10    | This PR (Fallback) | 0.049391028728263064 | 0.5183390069820635  | -1.0                 | 2   | 2   |
| 11    | Original           | 1.0                  | -1.0                | -1.0                 | 2   | 0   |
| 11    | This PR (Batched)  | 0.9999999999999998   | -1.0                | -1.0                 | 2   | 0   |
| 11    | This PR (Fallback) | 1.0                  | -1.0                | -1.0                 | 2   | 0   |
| 12    | Original           | 0.01292647202538677  | -1.0                | 0.683606077907549    | 2   | 1   |
| 12    | This PR (Batched)  | 0.012926472025411861 | -1.0                | 0.68360607790766     | 2   | 1   |
| 12    | This PR (Fallback) | 0.012926472025397429 | -1.0                | 0.6836060779076483   | 2   | 1   |
| 13    | Original           | -0.20230895590596554 | -1.0                | -0.6996612111149964  | 2   | 1   |
| 13    | This PR (Batched)  | -0.20230895590596643 | -1.0                | -0.6996612111149789  | 2   | 1   |
| 13    | This PR (Fallback) | -0.2023089559059673  | -1.0                | -0.6996612111149729  | 2   | 1   |

---

### 実験2: BBOB F6 (次元d=10)

連続変数のみを含むベンチマーク問題です。ベンチマークは[The blackbox optimization benchmarking (bbob) test suite \| OptunaHub](https://hub.optuna.org/benchmarks/bbob/)のものを利用しています。

**目的関数**
```python
import optunahub

bbob = optunahub.load_module("benchmarks/bbob")
f6 = bbob.Problem(function_id=6, dimension=10, instance_id=1)
```

#### 結果1. 実行時間と目的関数の最適値

| stats\mode   | This PR (Batched) | This PR (Fallback) | Original |
| ------------ | ----------------- | ------------------ | -------- |
| Runtime      | 3.83 sec          | 7.28 sec           | 6.93 sec |
| `best_value` | `237.65`          | `244.74`           | `244.74` |

#### 結果2. Trialの初期では、どの手法も最適解がほとんど一致する

| Trial | Mode               | x0                  | x1                  | x2                 | x3                 | x4                   | x5                   | x6                  | x7                  | x8                  | x9                  |
| ----- | ------------------ | ------------------- | ------------------- | ------------------ | ------------------ | -------------------- | -------------------- | ------------------- | ------------------- | ------------------- | ------------------- |
| 9     | Original           | -3.804057540616983  | 2.13244787222995    | 2.6078504861689744 | 0.6127719756949623 | 2.70967179954561     | -0.06204403635609257 | 0.22732829381994057 | -0.7245898164145037 | -4.745808732559048  | -3.9210857300669555 |
| 9     | This PR (Batched)  | -3.804057540616983  | 2.13244787222995    | 2.6078504861689744 | 0.6127719756949623 | 2.70967179954561     | -0.06204403635609257 | 0.22732829381994057 | -0.7245898164145037 | -4.745808732559048  | -3.9210857300669555 |
| 9     | This PR (Fallback) | -3.804057540616983  | 2.13244787222995    | 2.6078504861689744 | 0.6127719756949623 | 2.70967179954561     | -0.06204403635609257 | 0.22732829381994057 | -0.7245898164145037 | -4.745808732559048  | -3.9210857300669555 |
| 10    | Original           | -1.3521177949369183 | 1.884266535593106   | 4.70204332494208   | -4.686924455161294 | -4.6354277897651714  | -2.3837020295767584  | -2.54981063903124   | 3.9415049978118173  | -1.6310927165179656 | 4.931715192849079   |
| 10    | This PR (Batched)  | -1.3521177949369196 | 1.8842665355931043  | 4.702043324942078  | -4.686924455161291 | -4.63542778976517    | -2.3837020295767584  | -2.549810639031241  | 3.9415049978118155  | -1.631092716517966  | 4.931715192849079   |
| 10    | This PR (Fallback) | -1.3521177949369183 | 1.884266535593106   | 4.70204332494208   | -4.686924455161294 | -4.6354277897651714  | -2.3837020295767584  | -2.54981063903124   | 3.9415049978118173  | -1.6310927165179656 | 4.931715192849079   |
| 11    | Original           | -5.0                | -1.892109777045477  | 4.820142793954346  | 1.0039349678224934 | -5.0                 | -3.697589945694035   | -5.0                | 2.754227122961276   | -3.5443272688740404 | 4.676866136900413   |
| 11    | This PR (Batched)  | -5.0                | -1.8921097770447028 | 4.8201427939542185 | 1.003934967822202  | -5.0                 | -3.6975899456937182  | -5.0                | 2.754227122961492   | -3.5443272688737926 | 4.676866136900205   |
| 11    | This PR (Fallback) | -5.0                | -1.892109777045477  | 4.820142793954346  | 1.0039349678224934 | -5.0                 | -3.697589945694035   | -5.0                | 2.754227122961276   | -3.5443272688740404 | 4.676866136900413   |
| 12    | Original           | -0.8626904715067383 | 4.674728777076599   | 3.194588418725715  | -5.0               | -2.5966914635931295  | 1.0571190988778252   | -3.83263148929858   | 5.0                 | -2.138699815403413  | 4.712237655426524   |
| 12    | This PR (Batched)  | -0.8626904715060384 | 4.674728777076606   | 3.1945884187257327 | -5.0               | -2.5966914635929017  | 1.0571190988777293   | -3.8326314892982216 | 5.0                 | -2.138699815403209  | 4.712237655426639   |
| 12    | This PR (Fallback) | -0.8626904715067383 | 4.674728777076599   | 3.194588418725715  | -5.0               | -2.5966914635931295  | 1.0571190988778252   | -3.83263148929858   | 5.0                 | -2.138699815403413  | 4.712237655426524   |
| 13    | Original           | 1.7035918959822158  | 1.0887440040837895  | 4.13002366414174   | -5.0               | -0.16427108729306106 | -0.7893759531056306  | 0.1100295124370465  | 5.0                 | 3.3421161858848354  | 5.0                 |
| 13    | This PR (Batched)  | 1.70359189598196    | 1.0887440040837895  | 4.130023664141895  | -5.0               | -0.16427108729281947 | -0.7893759531055453  | 0.11002951243717529 | 5.0                 | 3.342116185885919   | 5.0                 |
| 13    | This PR (Fallback) | 1.7035918959822158  | 1.0887440040837895  | 4.13002366414174   | -5.0               | -0.16427108729306106 | -0.7893759531056306  | 0.1100295124370465  | 5.0                 | 3.3421161858848354  | 5.0                 |

---

## 考察 (Discussion)

- **実行時間**:
    - バッチ評価が機能した場合、実行時間が短縮されることが確認できました。特に、連続変数のみで構成される`f6`ベンチマークでは、約45%の高速化が達成されています。
    - フォールバックした場合の実行時間は、`Original`とほぼ同等であり、本実装によるオーバーヘッドが軽微であり、実装に不備がないと示唆されます。
- **精度**:
    - 100 Trials 後の`best_value`は、いずれの方式でも同等のオーダーに達しており、提案手法が最適化の精度を大きく損なっていないことが確認できました。
    - 初期のTrial（例: 9〜13）における提案パラメータは、数値誤差の範囲でほぼ一致しており、アルゴリズムの妥当性が示されています。